+++ src/initialize/createDotEnv.js
@@ -23,7 +23,11 @@ export const createDotEnv = async ({ authToken } = {}) => {
 
   const { source, dest } = getConfigValue('paths.templates.dotenv');
   const templateEnvFileContents = await readFile(source, { encoding: 'utf-8' });
-  const envFile = replaceTokens(envFileTokens, { authToken }, templateEnvFileContents);
+  const envFile = replaceTokens(
+    envFileTokens,
+    { authToken: authToken.trim() },
+    templateEnvFileContents
+  );
   logger.debug('saving .env file to: %s', dest);
   await outputFile(dest, envFile);
 };